+++ app/break.js
@@ -39,7 +39,7 @@ ipcRenderer.on('progress', (event, started, duration, strictMode, postpone, post
       closeElement.style.display =
         Utils.canSkip(strictMode, postpone, passedPercent, postponePercent) ? 'flex' : 'none'
       progress.value = (100 - passedPercent) * progress.max / 100
-      progressTime.innerHTML = Utils.formatRemaining(Math.trunc((duration - Date.now() + started) / 1000))
+      progressTime.innerHTML = Utils.formatTimeRemaining(Math.trunc(duration - Date.now() + started))
     }
   }, 100)
 })