+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - better style for progress
 - rework Contributor preferences
+- synchronized tray status messages (tooltip, menu)
+- better format remaining time
 
 ## [0.99.1] - 2020-04-09
 ### Changed